Moving iree/base/arena.h to iree/hal/vulkan/util/. This will allow iree/hal/local/arena.h to be swapped in its place.
