+++ iree/hal/vulkan/BUILD
@@ -90,7 +90,6 @@ cc_library(
     deps = [
         ":dynamic_symbols",
         "//iree/base:api",
-        "//iree/base:arena",
         "//iree/base:core_headers",
         "//iree/base:flatcc",
         "//iree/base:logging",
@@ -98,6 +97,7 @@ cc_library(
         "//iree/base:synchronization",
         "//iree/base:tracing",
         "//iree/hal:api",
+        "//iree/hal/vulkan/util:arena",
         "//iree/hal/vulkan/util:intrusive_list",
         "//iree/hal/vulkan/util:ref_ptr",
         "//iree/schemas:spirv_executable_def_c_fbs",
