+++ iree/hal/vulkan/util/arena_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iree/base/arena.h"
+#include "iree/hal/vulkan/util/arena.h"
 
 #include "iree/testing/gtest.h"
